add menu and update ui
