+++ myWeather/Extensions/Unit+Extensions.swift
@@ -1,0 +1,10 @@
+//
+//  Unit+Extensions.swift
+//  myWeather
+//
+//  Created by eleazar.martinez on 9/30/22.
+//
+
+import Foundation
+
+
