--- myWeather/Extensions/View+Extensions.swift
@@ -1,17 +1,0 @@
-//
-//  View+Extensions.swift
-//  WeatherAppSwiftUI
-//
-//  Created by Mohammad Azam on 3/5/21.
-//
-
-import Foundation
-import SwiftUI
-
-extension View {
-    
-    func embedInNavigationView() -> some View {
-        return NavigationView { self }
-    }
-    
-}
